# parser/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dmarc-parser API",
    "version": "2.0.0",
    "description": "REST API of the DMARC aggregate report parser & ingestion service.\n\n### Authentication\nEvery `/api/v1/*` route (except `/api/v1/openapi.json`) requires an API\nkey from `PARSER_API_KEYS`, sent either as `Authorization: Bearer <key>`\nor `X-API-Key: <key>`. Keys carry scopes from `{read, ingest, admin}`;\neach operation documents its required scope via `x-required-scope`.\nMissing/invalid key \u2192 401; key without the required scope \u2192 403.\n\n### Rate limiting\nPer-key token bucket (`PARSER_RATE_LIMIT_RPS`, default 10 req/s,\nburst 3x). Exceeding it returns 429 with a `Retry-After` header.\n\n### Conventions\nAll timestamps are UTC, RFC 3339. Date-only values accepted by `since` /\n`until` parameters use `YYYY-MM-DD`. Errors are JSON\n`{\"error\": \"message\"}`.\n"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8081",
      "description": "Host loopback"
    },
    {
      "url": "http://dmarcparser:8080",
      "description": "From containers on the mxsentinel_default / dmarcparser_default networks"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/sources": {
      "get": {
        "tags": [
          "intel"
        ],
        "summary": "Threat list of sending IPs across all domains",
        "description": "Per-IP totals, failed message counts, targeted domain count and ip_meta enrichment, ordered by failing messages descending.",
        "operationId": "listSources",
        "x-required-scope": "read",
        "parameters": [
          {
            "$ref": "#/components/parameters/domain"
          },
          {
            "name": "min_failed",
            "in": "query",
            "description": "Only IPs with at least this many failing messages",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "$ref": "#/components/parameters/since"
          },
          {
            "$ref": "#/components/parameters/limit1000"
          }
        ],
        "responses": {
          "200": {
            "description": "Source list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sources": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Source"
                      }
                    },
                    "limit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/ips/{ip}": {
      "get": {
        "tags": [
          "intel"
        ],
        "summary": "Cross-report activity and metadata for one IP",
        "operationId": "getIP",
        "x-required-scope": "read",
        "parameters": [
          {
            "name": "ip",
            "in": "path",
            "required": true,
            "description": "IPv4 or IPv6 address (v4-mapped v6 is unmapped)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "IP drill-down",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IPDetail"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
    "/api/v1/domains/{domain}/sources": {
      "get": {
        "tags": [
          "intel"
        ],
        "summary": "Sending sources of one domain",
        "operationId": "getDomainSources",
        "x-required-scope": "read",
        "parameters": [
          {
            "$ref": "#/components/parameters/domainPath"
          },
          {
            "name": "unacked",
            "in": "query",
            "description": "Only sources not yet acknowledged",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Domain sources",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "domain": {
                      "type": "string"
                    },
                    "unacked_only": {
                      "type": "boolean"
                    },
                    "sources": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DomainSource"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
    "/api/v1/sources/ack": {
      "post": {
        "tags": [
          "intel"
        ],
        "summary": "Acknowledge a (domain, ip) source",
        "description": "Acked sources stop being reported as new senders.",
        "operationId": "ackSource",
        "x-required-scope": "admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "domain",
                  "ip"
                ],
                "properties": {
                  "domain": {
                    "type": "string"
                  },
                  "ip": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Acknowledged",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "const": "acked"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key from PARSER_API_KEYS as a bearer token. Key scopes: read, ingest, admin (see each operation's x-required-scope)."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Same API key via the X-API-Key header."
      }
    },
    "parameters": {
      "domain": {
        "name": "domain",
        "in": "query",
        "description": "Filter by policy-published domain",
        "schema": {
          "type": "string"
        }
      },
...
      "since": {
        "name": "since",
        "in": "query",
        "description": "Lower time bound (RFC 3339 or YYYY-MM-DD)",
        "schema": {
          "type": "string"
        }
      },
...
      "limit1000": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 1000,
          "default": 100
        }
      },
...
      "domainPath": {
        "name": "domain",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "API key lacks the required scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Per-key rate limit exceeded (see Retry-After)",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying",
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "Source": {
        "type": "object",
        "properties": {
          "ip": {
            "type": "string"
          },
          "msgs_total": {
            "type": "integer",
            "format": "int64"
          },
          "msgs_failed": {
            "type": "integer",
            "format": "int64"
          },
          "domains": {
            "type": "integer",
            "format": "int64",
            "description": "Distinct domains targeted"
          },
          "first_seen": {
            "type": "string",
            "format": "date-time"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time"
          },
          "ptr": {
            "type": [
              "string",
              "null"
            ]
          },
          "asn": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "as_org": {
            "type": [
              "string",
              "null"
            ]
          },
          "country": {
            "type": [
              "string",
              "null"
            ]
          },
          "sender_class": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "IPMeta": {
        "type": "object",
        "properties": {
          "ip": {
            "type": "string"
          },
          "ptr": {
            "type": [
              "string",
              "null"
            ]
          },
          "asn": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "as_org": {
            "type": [
              "string",
              "null"
            ]
          },
          "country": {
            "type": [
              "string",
              "null"
            ]
          },
          "sender_class": {
            "type": [
              "string",
              "null"
            ]
          },
          "first_seen": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "refreshed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "DomainSource": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string"
          },
          "ip": {
            "type": "string"
          },
          "first_seen": {
            "type": "string",
            "format": "date-time"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time"
          },
          "first_serial": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "msg_count": {
            "type": "integer",
            "format": "int64"
          },
          "aligned_count": {
            "type": "integer",
            "format": "int64"
          },
          "acked": {
            "type": "boolean"
          },
          "ptr": {
            "type": [
              "string",
              "null"
            ]
          },
          "asn": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "as_org": {
            "type": [
              "string",
              "null"
            ]
          },
          "country": {
            "type": [
              "string",
              "null"
            ]
          },
          "sender_class": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "IPReportActivity": {
        "type": "object",
        "properties": {
          "serial": {
            "type": "integer",
            "format": "int64"
          },
          "domain": {
            "type": "string"
          },
          "org": {
            "type": "string"
          },
          "date_begin": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "date_end": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "messages": {
            "type": "integer",
            "format": "int64"
          },
          "disposition": {
            "type": [
              "string",
              "null"
            ]
          },
          "dkim_align": {
            "type": "string"
          },
          "spf_align": {
            "type": "string"
          },
          "header_from": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "IPDetail": {
        "type": "object",
        "properties": {
          "ip": {
            "type": "string"
          },
          "meta": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/IPMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "domains": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DomainSource"
            }
          },
          "reports": {
            "type": "array",
            "description": "Latest 100 report records featuring this IP",
            "items": {
              "$ref": "#/components/schemas/IPReportActivity"
            }
          }
        }
      },
...
    }
  }
}